Add new test
